requirements.txt
python == 3.6
django == 2.7
djangorestframework
